collectd: disable nut plugin
Temporarily disable nut plugin by marking it BROKEN.
Apparently the upstream collectd plugin is not yet compatible
with nut 2.8.0.

I tried the upstream PR 4043, but it does not help, as it
apparently requires at least one specific UPS type to be
defined/compiled.

        nut . . . . . . . . . no (required data types for
            NUT API were not detected) (dependency error)

If there is a requirement to specify some UPS data type when
compiling nut, it might be problematic for the generic OpenWrt
buildbot context.

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>