net-snmp: change reload and config handling
Add the generated snmpd config file /var/run/snmpd.conf to the snmpe start
command with option option '-c $CONFIGFILE' and to net search for
configuration files with the start option '-C'. So we know exactly which
configuration the service uses. Until now it was not clear, because it
took the first one it found.

Also, this commit removed the `procd_set_param file $CONFIGFILE`. This
option triggers a reload if the file has been rewritten or modified.
However, this does not make sense as this file is rewritten with every uci
change. This would mean that the service is restarted twice.

Since the snmdp config file is now specified explicitly, the generated
symbolic link can be omitted during installation as well.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>