gdbm: configure with `--without-readline`
By default, gdbm configures/builds with readline and curses support
if they exist.

This can cause race conditions when compiling gdbm and ncurses in
parallel, as gdbm may try to link to ncurses when it doesn't exist.

This commit forces gdbm to skip using readline/ncurses, since it's
unlikely that anybody is using the line-editing feature of gdbm.

See [gdbm's README][1] for more info.

[1]: https://git.gnu.org.ua/gdbm.git/tree/README?h=v1.21#n50

Signed-off-by: Alois Klink <alois@aloisklink.com>