prometheus-node-exporter-lua: wifi: Add a metric for the number of stations
The number of stations is already provided by the wifi stations
collector. But this collector also provides per station metrics which
can create an unbonded number of time series in the prometheus
database. Having such metrics can also be a privacy issues depending
on who can access the monitoring data.

But the number of connected stations is still an interesting metric,
especially if you have many access points. This patch add the
`wifi_network_stations_total` metric to the main wifi collector to
fill this gap.

Signed-off-by: Alban Bedel <albeu@free.fr>