libgpg-error: copy gpg-error.pc to staging directory
Packages like libassuan and libksba would not compile after updating
libgpg-error to 1.46. This was because gpg-error.m4 (and thus the affected
packages' configure scripts) relies on gpgrt-config, which in turn needs
access to gpg-error.pc. This modifies the libgpg-error build process to
copy gpg-error.pc to OpenWrt's staging directory, so that it is available
for subsequent dependent package builds.

Fixes: https://github.com/openwrt/packages/issues/19880

Signed-off-by: W. Michael Petullo <mike@flyn.org>