mwan3: make mwan3.user executable
This is a fix for the the following change:
https://github.com/openwrt/packages/commit/3d824ea288d907a31729c3629e884ea122c30da0

Before the change, it was only possible to execute a shell script. To
remove this restriction, a binary or other script language can now also
be used for 'mwan3.user'. Unfortunately, the old shell script was not
executable for older mwan3 version. During a sysupgrade with config transfer,
this 'mwan3.user' script could not be executed for newer mwan3 versions.
To fix this, the calling script checks whether the 'mwan3.user' is executable,
and if not, this executable bit is now set.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 0434407428b25ae5309525f060964218034b86c5)