zyxel-reset: add package
Add the zyxel-reset package.

This package allows to trigger a factory-reset for ZyXEL devices by
sending a magic LLDP package while the device-to-reset is booting.

This is useful for remote-resetting a ZyXEL device running stock
firmware connected to a switch using OpenWrt. It also allows to reset
devices which do not have a reset-button such as the NWA55AXE.

Signed-off-by: David Bauer <mail@david-bauer.net>