stubby: create appdata_dir if it doesn't exist
From https://github.com/getdnsapi/stubby/blob/b53867d1ec467822dbd9e1a0ad6598f84a9f5ed7/stubby.yml.example#L150:

"Stubby tries to fetch and validate the DNSSEC root trust anchor on the fly when needed (zero-configuration DNSSEC), but only if it can store them somewhere. The default location to store these files is the `.getdns` subdirectory in the user's home directory on Unixes, and the %appdata%\getdns directory on Windows. If there is no home directory, or the required subdirectory could not be created (or is not present), Stubby will fall back to the current working directory to try to store the trust-anchor files.

"When stubby runs as a special system-level user without a home directory however (such as in setups using systemd), it is recommended that an explicit location for storing the trust-anchor files is provided that is writable (and readable) by that special system user."

Previously, stubby's makefile and initscript never created the `appdata_dir`. Therefore, create the `appdata_dir` if it doesn't exist.

And while we're at it, run stubby with the `no_new_privs` bit set.

Signed-off-by: Aquila Cooper <aquila@cpr.is>