xxhash: Fix pkg-config metadata file
Currently, Build/InstallDev installs libxxhash.pc.in, which is the
template for libxxhash.pc and does not contain the correct information.
(pkg-config also does not recognize this file name.)

This uses PKG_INSTALL:=1 so that libxxhash.pc is generated (by 'make
install').

This also removes the need to override Build/Prepare by setting the
correct command in PKG_UNPACK.

Signed-off-by: Jeffery To <jeffery.to@gmail.com>