python-editables: Add new host-only package
From the README:

A Python library for creating "editable wheels"

This library supports the building of wheels which, when installed, will
expose packages in a local directory on sys.path in "editable mode". In
other words, changes to the package source will be reflected in the
package visible to Python, without needing a reinstall.

Signed-off-by: Jeffery To <jeffery.to@gmail.com>