zabbix: add patch for gnutls version handling
Zabbix verifies the version of gnutls by search in the header gnutls.h.
This is done with 'cat' and 'egrep'. The problem here is now that the
preprocess definition did changed in gnutls.h, so the regex does not
match anymore. The following error message acourse in the log.

configure: error: GnuTLS library version requirement not met (>= 3.1.18)

To fix this the regex must be correced with this commit, so the version
check does work again

Signed-off-by: Florian Eckert <fe@dev.tdt.de>