znc: add push external module
Add push external module. This is very useful for an IRC Bounder as this
module permits to register various services and receive a push
notification on the registered service.

One example is attaching a telegram bot and receive notification on your
phone when an user tags you in one of the connected channels.

Bump and drop AUTORELEASE from PKG_RELEASE since we are adding a new
module.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>