modemmanager: block restart only in case of wrong credentials
It may happen that a mobile router is powered on while out of network
coverage. In this case it should keep trying to connect but currently
restart is blocked if the initial attempt to connect fails.
Fix this by blocking restart only if the PIN code or connection
credentials are wrong or missing.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>