ruby: build hostpkg with rpath
Compile currently fails in case host system doesn't provide a fitting
libyaml, which is unnecessary because ruby already depends yaml/host, so
the lib is there, ruby hostpkg just can't find it.

This commit adds an rpath to the host LDFLAGS to resolve this. An
alternative would be to build a static ruby hostpkg, but why waste the
space.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>