mwan3: fix ipv6 tunnels assigned the wrong mark
The mark of underlying ipv4 tunnel connection is assigned to all incoming ipv6 packets. This breaks mwan3 ipv6 routing. Attempt to fix this by clearing the mark in incoming packets on wan interfaces. Do not touch outgoing packets to make sure that tracking and "mwan3 use" command works as expected.

Signed-off-by: Anna Tikhomirova <vamp@vampik.ru>