python-packages: Remove __PYVENV_LAUNCHER__ environment variable
Setting __PYVENV_LAUNCHER__ for Python package builds was added in
a91a992abb1fe9b7c4e4e5d11532dbf18c5b213d, but neither the commit message
nor the pull request[1] explain its purpose in detail.

My guess is this was done to set the shebang for installed Python
scripts. We now have a Makefile recipe to set the shebang, so it would
be unnecessary to set this variable for this purpose.

It appears that Python 3.11 has changed the handling of this (internal)
environment variable, and setting it appears to be causing build errors
for all Python packages.

This removes setting __PYVENV_LAUNCHER__ for Python package builds.

[1]: https://github.com/openwrt/packages/pull/525

Fixes: https://github.com/openwrt/packages/issues/21162
Signed-off-by: Jeffery To <jeffery.to@gmail.com>