cudy-bdinfo: add package
This program can be used to obtain information stored on the bdinfo
parition found on routers from Shenzhen Cudy Technology.

While this tool is not necessary for operation with OpenWrt, it can be
helpful to decrypt information generated by the stock firmware as well
as to obtain additional device-specific information.

Signed-off-by: David Bauer <mail@david-bauer.net>