modemmanager: if an alias name is used do not check sysfs path
If an alias name is used for the modem, then a check if the device exists
in sysfs does not work. To fix this remove the check if the sysfs device
exists. The protocoll handler already checks if the modem is responsible
for this device on the next line.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit fd63e400ba1c2cf7242517ebb5675bfdcd22fc3f)