modemmanager: do not set proto_notify_error on teardown
If on teardown the 'proto_notify_error' is set to 'MM_TEARDOWN_IN_PROGRESS',
then an error which is set on 'setup' is not visible in the ubus
network.interface.<iface> status output.

{
  "up": false,
  "pending": false,
  "available": true,
  "autostart": false,
  "dynamic": false,
  "proto": "modemmanager",
  "data": {
  },
  "errors": [
    {
      "subsystem": "dualsim",
      "code": "MM_TEARDOWN_IN_PROGRESS"
    }
  ]
}

It alway shows the code 'MM_TEARDWON_IN_PROGRESS'!

By removing the line 'proto_notify_error "${interface}" MM_TEARDOWN_IN_PROGRESS'
in teardown, the last error is show in the proto stack from setup.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>