collectd: unify configure args handling for collectd-mod-smart
If 'collectd-mod-smart' is not enabled, the use of lubudev does not need
to be explicitly disabled. This library could be a dependency for other
modules in the future. If this is explicitly disabled, then the other
module cannot be built.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 1aac8ada877d282c593895e7ce5b31134cb08aac)