vnstat2: update to version 2.11
This release breaks the noexit patch, because the code for removing old
now returns an error when no interfaces are configured. As it is run on
startup, the daemon exits in this case. To avoid this, add an additional
check so an error is only returned in an actual error case.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>
(cherry picked from commit 98719797074637cecdff6a0cdb782a1e807a13e9)